+++ near-contract-standards/src/non_fungible_token/approval/approval_receiver.rs
