+++ near-contract-standards/src/non_fungible_token/core/mod.rs
@@ -1,3 +1,11 @@
+mod core_impl;
+mod receiver;
+mod resolver;
+
+pub use self::core_impl::*;
+pub use self::receiver::*;
+pub use self::resolver::*;
+
 use crate::non_fungible_token::metadata::TokenMetadata;
 use crate::non_fungible_token::token::{Token, TokenId};
 use near_sdk::json_types::ValidAccountId;
