+++ near-contract-standards/src/non_fungible_token/core/receiver.rs
