+++ near-contract-standards/src/non_fungible_token/enumeration/enumeration_impl.rs
@@ -1,6 +1,6 @@
-use crate::non_fungible_token::core_impl::NonFungibleToken;
-use crate::non_fungible_token::enumeration::NonFungibleTokenEnumeration;
+use super::NonFungibleTokenEnumeration;
 use crate::non_fungible_token::token::Token;
+use crate::non_fungible_token::NonFungibleToken;
 use near_sdk::json_types::{ValidAccountId, U128};
 use near_sdk::AccountId;
